OneForAll_Smart_Glove/PC_Scripts/requirements.txt
pyserial==3.5
pyautogui==0.9.52